Implemented Auth service

+++ src/main/java/com/app/ECommerceWebApp/controllers/CartController.java
@@ -4,9 +4,10 @@ import com.app.ECommerceWebApp.dtos.cartItemDtos.AddCartItemDTO;
 import com.app.ECommerceWebApp.dtos.cartItemDtos.RemoveCartItemDTO;
 import com.app.ECommerceWebApp.dtos.cartItemDtos.UpdateQuantityDTO;
 import com.app.ECommerceWebApp.exceptions.cartExceptions.CartNotExistsException;
+import com.app.ECommerceWebApp.exceptions.userExceptions.InvalidTokenException;
 import com.app.ECommerceWebApp.models.Cart;
-import com.app.ECommerceWebApp.services.CartItemService;
-import com.app.ECommerceWebApp.services.CartService;
+import com.app.ECommerceWebApp.services.cart.CartService;
+import com.app.ECommerceWebApp.utils.AuthUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,14 +17,20 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/cart")
 public class CartController {
     private CartService cartService;
+    private AuthUtils authUtils;
 
     @Autowired
-    public CartController(CartService cartService) {
+    public CartController(CartService cartService, AuthUtils authUtils) {
         this.cartService = cartService;
+        this.authUtils = authUtils;
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Cart> viewCart(@PathVariable long id){
+    public ResponseEntity<Cart> viewCart(@PathVariable long id, @RequestHeader String value){
+        if(!authUtils.validate_token(value)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             Cart cart = this.cartService.viewCart(id);
             return new ResponseEntity<>(cart, HttpStatus.OK);
@@ -33,7 +40,11 @@ public class CartController {
     }
 
     @PostMapping("/addCartItem")
-    public ResponseEntity<Cart> addCartItem(@RequestBody AddCartItemDTO cartItemDTO){
+    public ResponseEntity<Cart> addCartItem(@RequestBody AddCartItemDTO cartItemDTO, @RequestHeader String value){
+        if(!authUtils.validate_token(value)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             Cart cart = this.cartService.addCartItem(cartItemDTO.getCart_id(), cartItemDTO.getProduct_id(), cartItemDTO.getQuantity());
             return new ResponseEntity<>(cart, HttpStatus.CREATED);
@@ -43,7 +54,11 @@ public class CartController {
     }
 
     @DeleteMapping("/removeCartItem")
-    public ResponseEntity<Cart> removeCartItem(@RequestBody RemoveCartItemDTO removeCartItemDTO){
+    public ResponseEntity<Cart> removeCartItem(@RequestBody RemoveCartItemDTO removeCartItemDTO, @RequestHeader String value){
+        if(!authUtils.validate_token(value)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             Cart cart = this.cartService.removeCartItem(removeCartItemDTO.getCart_id(), removeCartItemDTO.getProduct_id());
             return new ResponseEntity<>(cart, HttpStatus.NO_CONTENT);
@@ -54,7 +69,11 @@ public class CartController {
     }
 
     @PatchMapping("/updateQuantity")
-    public ResponseEntity<Cart> updateQuantity(@RequestBody UpdateQuantityDTO updateQuantityDTO){
+    public ResponseEntity<Cart> updateQuantity(@RequestBody UpdateQuantityDTO updateQuantityDTO, @RequestHeader String value){
+        if(!authUtils.validate_token(value)){
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             Cart cart = this.cartService.updateQuantity(updateQuantityDTO.getCart_id(), updateQuantityDTO.getProduct_id(), updateQuantityDTO.getQuantity());
             return new ResponseEntity<>(cart, HttpStatus.CREATED);
